carshop_api/models.py
# Create your models here.
from django.db.models.signals import post_save
from django.dispatch import receiver
from rest_framework.authtoken.admin import User
from rest_framework.authtoken.models import Token

