requirements.txt
django>=5.0,<5.1
gunicorn>=21.2,<22.0
dj-database-url>=2.0,<3.0
whitenoise[brotli]>=6.0,<7.0
pytest-django
django-allauth
django-crispy-forms
django-redis
redis
djangorestframework
drf-yasg
djangorestframework-simplejwt

# Uncomment these lines to use a Postgres database. Both are needed, since in production
# (which uses Linux) we want to install from source, so that security updates from the
# underlying Heroku stack image are picked up automatically, thanks to dynamic linking.
# On other platforms/in development, the precompiled binary package is used instead, to
# speed up installation and avoid errors from missing libraries/headers.
psycopg[c]; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"
